+++ nn_meter/builder/kernel_predictor_builder/data_sampler/utils.py
@@ -111,41 +111,3 @@ def get_sampler_for_kernel(kernel_type, sample_num, sampling_mode, configs = Non
 
 def list_kernels():
     return list(__BUILTIN_KERNELS__.keys()) + ["* " + item for item in list(__REG_KERNELS__.keys())]
-
-
-def inverse_transform_sampling(data, n_bins = 40, n_samples = 1000):
-    ''' calculate inversed cdf, for sampling by possibility
-    '''
-    import scipy.interpolate as interpolate
-    hist, bin_edges = np.histogram(data, bins=n_bins, density=True)
-    cum_values = np.zeros(bin_edges.shape)
-    cum_values[1:] = np.cumsum(hist*np.diff(bin_edges))
-    inv_cdf = interpolate.interp1d(cum_values, bin_edges)
-    r = np.random.rand(n_samples)
-    data = inv_cdf(r)
-    ndata = [int(x) for x in data]
-    return ndata
-
-
-def sample_based_on_distribution(data, count):
-    ''' use data to calculate a inversed cdf, and sample `count` data from such distribution
-    '''
-    return inverse_transform_sampling(data, n_samples=count)
-
-
-def data_validation(data, cdata):
-    ''' convert sampled data to valid configuration, e.g.,: kernel size = 1, 3, 5, 7
-
-    @params:
-    data: the origin data value.
-    cdata: valid configuration value.
-    '''
-    newlist = []
-    for da in cdata:
-        value = [abs(da - x) for x in data]
-        newlist.append(value)
-
-    newlist = list(np.asarray(newlist).T)    
-    cda = [list(d).index(min(d)) for d in newlist]
-    redata = [cdata[x] for x in cda]
-    return redata
